Update blit to 0.4

+++ src/terrain.rs
@@ -3,12 +3,12 @@ use line_drawing::Bresenham;
 pub struct Terrain {
     pub buffer: Vec<u32>,
 
-    width: i32,
-    height: i32
+    width: usize,
+    height: usize
 }
 
 impl Terrain {
-    pub fn new(size: (i32, i32)) -> Self {
+    pub fn new(size: (usize, usize)) -> Self {
         Terrain {
             buffer: vec![0xFFFF00FF; (size.0 * size.1) as usize],
 
@@ -17,7 +17,7 @@ impl Terrain {
         }
     }
 
-    pub fn size(&self) -> (i32, i32) {
+    pub fn size(&self) -> (usize, usize) {
         (self.width, self.height)
     }
 
@@ -25,11 +25,11 @@ impl Terrain {
         let (width, height) = self.size();
 
         for pos in Bresenham::new(start, end) {
-            if pos.0 < 0 || pos.1 < 0 || pos.0 >= width || pos.1 >= height {
+            if pos.0 < 0 || pos.1 < 0 || pos.0 as usize >= width || pos.1 as usize >= height {
                 continue;
             }
 
-            let index = (pos.0 + pos.1 * width) as usize;
+            let index = pos.0 as usize + pos.1 as usize * width;
             if (self.buffer[index] & 0xFFFFFF) != 0xFF00FF {
                 return Some(pos);
             }
